+++ src/components/Alert/alert.css
@@ -1,0 +1,54 @@
+@keyframes fadeOut {
+  0%,
+  100% {
+    opacity: 0;
+  }
+  10%,
+  90% {
+    opacity: 1;
+  }
+}
+
+.alert-container {
+  position: fixed;
+  left: 50%;
+  bottom: 5%;
+
+  transform: translate(-50%, -50%);
+
+  min-width: 40%;
+  max-width: 70%;
+  height: 38px;
+
+  color: #000;
+  background: #ffffff;
+
+  box-shadow: 1px 1px 14.6px 3px #ea0202;
+  border-radius: 7px;
+
+  opacity: 0.9;
+  padding: 0 1rem;
+
+  font-size: 1.25rem;
+  font-weight: 700;
+
+  animation: fadeOut 3s ease-in-out 0s infinite normal none running;
+}
+
+.alert-content {
+  height: 100%;
+  display: flex;
+  gap: 0.5rem;
+  justify-content: center;
+  align-items: center;
+}
+
+.alert-text {
+  overflow: hidden;
+  text-overflow: ellipsis;
+  -webkit-line-clamp: 1;
+  -webkit-box-orient: vertical;
+
+  white-space: nowrap;
+  text-align: center;
+}
