+++ src/components/ErrorPage/error-page.css
@@ -1,0 +1,19 @@
+.error {
+  position: absolute;
+  left: 50%;
+  top: 50%;
+
+  filter: grayscale(100%);
+
+  transform: translate(-50%, -50%);
+}
+
+.status-number {
+  margin-right: 1rem;
+  font-size: 2rem;
+  color: tomato;
+}
+
+.error-message {
+  font-size: 2rem;
+}
