+++ src/components/Title/title.css
@@ -1,0 +1,8 @@
+.title {
+  cursor: pointer;
+  user-select: none;
+  font-size: 2rem;
+  font-weight: bold;
+  letter-spacing: -0.1rem;
+  color: #f33f3f;
+}
